fix(guard): split the emergency-safe hook floor below the 500-line ratchet
Keep Cursor fail-closed hooks from growing new oversized modules while the degraded inspection classifier stays independently reviewable.
