+++ docs/guard/contracts/hook-data-plane-ownership.v2.json
@@ -278,6 +278,7 @@
         "src/codex_plugin_scanner/guard/daemon/hook_worker_native.py",
         "src/codex_plugin_scanner/guard/daemon/hook_worker_responses.py",
         "src/codex_plugin_scanner/guard/daemon/hook_availability_policy.py",
+        "src/codex_plugin_scanner/guard/daemon/hook_availability_floor.py",
         "src/codex_plugin_scanner/guard/native_approval_bridge.py",
         "src/codex_plugin_scanner/guard/native_approval_bridge_v3.py",
         "src/codex_plugin_scanner/guard/native_approval_errors.py",
